backend/requirements.txt
camelot_py==1.0.0
fastapi[standard]==0.116.0
mistralai==1.9.2
numpy==2.3.1
opencv_python_headless==4.11.0.86
pandas==2.3.1
pdfplumber==0.11.7
pydantic==2.11.7
pymupdf==1.26.3
pytesseract==0.3.13
scipy==1.16.3
openai==2.8.1
